move base fee computations to a submodule
